+++ WebApplication/src/components/StudyItem.vue
@@ -3,6 +3,7 @@ import SeriesList from "./SeriesList.vue"
 import StudyDetails from "./StudyDetails.vue";
 import { mapState } from "vuex"
 import bootstrap from "bootstrap/dist/js/bootstrap.bundle.min.js"
+import api from "../orthancApi";
 
 export default {
     props: ["studyId"],
@@ -61,6 +62,14 @@ export default {
         onDeletedStudy(studyId) {
             this.$emit("deletedStudy", this.studyId);
         },
+        async onLabelsUpdated(studyId) {
+            this.$store.dispatch('studies/reloadStudy', {
+                'studyId': studyId,
+                'study': await api.getStudy(studyId)
+            })
+            const study = this.studies.filter(s => s["ID"] == this.studyId)[0];
+            this.fields = study;
+        },
         onSelectedStudy() {
             this.selected = true;
         },
@@ -75,10 +84,17 @@ export default {
         ...mapState({
             uiOptions: state => state.configuration.uiOptions,
             studies: state => state.studies.studies,
+            allLabels: state => state.labels.allLabels
         }),
         modalitiesInStudyForDisplay() {
             return this.fields.RequestedTags.ModalitiesInStudy.split('\\').join(',');
         },
+        showLabels() {
+            return !this.expanded && ((this.allLabels && this.allLabels.length > 0));
+        },
+        hasLabels() {
+            return this.fields && this.fields.Labels && this.fields.Labels.length > 0;
+        }
     },
     components: { SeriesList, StudyDetails }
 }
@@ -87,7 +103,7 @@ export default {
 
 <template>
     <tbody>
-        <tr v-if="loaded" :class="{ 'study-row-collapsed': !expanded, 'study-row-expanded': expanded }">
+        <tr v-if="loaded" :class="{ 'study-row-collapsed': !expanded, 'study-row-expanded': expanded, 'study-row-show-labels': showLabels }">
             <td>
                 <div class="form-check">
                     <input class="form-check-input" type="checkbox" v-model="selected" @click="clickedSelect">
@@ -99,27 +115,37 @@ export default {
                 <span v-if="columnTag == 'StudyDate'" data-bs-toggle="tooltip"
                     v-bind:title="fields.MainDicomTags.StudyDate">{{ fields.MainDicomTags.StudyDate }}
                 </span>
-                <span v-if="columnTag == 'AccessionNumber'" data-bs-toggle="tooltip"
+                <span v-else-if="columnTag == 'AccessionNumber'" data-bs-toggle="tooltip"
                     v-bind:title="fields.MainDicomTags.AccessionNumber">{{ fields.MainDicomTags.AccessionNumber }}
                 </span>
-                <span v-if="columnTag == 'PatientID'" data-bs-toggle="tooltip"
+                <span v-else-if="columnTag == 'PatientID'" data-bs-toggle="tooltip"
                     v-bind:title="fields.PatientMainDicomTags.PatientID">{{ fields.PatientMainDicomTags.PatientID }}
                 </span>
-                <span v-if="columnTag == 'PatientName'" data-bs-toggle="tooltip"
+                <span v-else-if="columnTag == 'PatientName'" data-bs-toggle="tooltip"
                     v-bind:title="fields.PatientMainDicomTags.PatientName">{{ fields.PatientMainDicomTags.PatientName }}
                 </span>
-                <span v-if="columnTag == 'PatientBirthDate'" data-bs-toggle="tooltip"
+                <span v-else-if="columnTag == 'PatientBirthDate'" data-bs-toggle="tooltip"
                     v-bind:title="fields.PatientMainDicomTags.PatientBirthDate">{{
                         fields.PatientMainDicomTags.PatientBirthDate }}
                 </span>
-                <span v-if="columnTag == 'StudyDescription'" data-bs-toggle="tooltip"
+                <span v-else-if="columnTag == 'StudyDescription'" data-bs-toggle="tooltip"
                     v-bind:title="fields.MainDicomTags.StudyDescription">{{ fields.MainDicomTags.StudyDescription }}
                 </span>
-                <span v-if="columnTag == 'modalities'" data-bs-toggle="tooltip" v-bind:title="modalitiesInStudyForDisplay">{{
-                    modalitiesInStudyForDisplay }}
+                <span v-else-if="columnTag == 'modalities'" data-bs-toggle="tooltip"
+                    v-bind:title="modalitiesInStudyForDisplay">{{
+                        modalitiesInStudyForDisplay }}
                 </span>
-                <span v-if="columnTag == 'seriesCount'">{{ fields.Series.length }}
+                <span v-else-if="columnTag == 'seriesCount'">{{ fields.Series.length }}
                 </span>
+                <span v-else>{{ fields.MainDicomTags[columnTag] }}
+                </span>
+            </td>
+        </tr>
+        <tr v-show="showLabels">
+            <td></td>
+            <td colspan="100%" class="label-row">
+                <span v-for="label in fields.Labels" :key="label" class="label badge bg-info">{{ label }}</span>
+                <span v-if="!hasLabels">&nbsp;</span>
             </td>
         </tr>
         <tr v-show="loaded" class="collapse"
@@ -127,25 +153,32 @@ export default {
             v-bind:id="'study-details-' + this.studyId" ref="study-collapsible-details">
             <td v-if="loaded && expanded" colspan="100">
                 <StudyDetails :studyId="this.studyId" :studyMainDicomTags="this.fields.MainDicomTags"
-                    :patientMainDicomTags="this.fields.PatientMainDicomTags" @deletedStudy="onDeletedStudy"></StudyDetails>
+                    :patientMainDicomTags="this.fields.PatientMainDicomTags" :labels="this.fields.Labels" @deletedStudy="onDeletedStudy" @studyLabelsUpdated="onLabelsUpdated"></StudyDetails>
             </td>
         </tr>
     </tbody>
 </template>
 
 <style scoped>
-.study-row-collapsed {}
+.study-row-collapsed {
+    border-top-width: 1px;
+    border-color: #ddd;
+}
 
 .study-row-expanded {
     background-color: var(--study-selected-color);
     font-weight: 700;
 
-    border-top: 2px !important;
+    border-top: 3px !important;
     border-style: solid !important;
     border-color: black !important;
 }
 
 .study-row-expanded>:first-child {
+    border-bottom: 5px !important;
+}
+
+.study-row-show-labels {
     border-bottom: 0px !important;
 }
 
@@ -165,8 +198,18 @@ export default {
     background-color: var(--study-selected-color);
 
     border-top: 0px !important;
-    border-bottom: 2px !important;
+    border-bottom: 3px !important;
     border-style: solid !important;
     border-color: black !important;
+}
+
+.label {
+    margin-left: 2px;
+    margin-left: 2px;
+}
+
+.label-row {
+    border-top: 0px !important;
+    border-bottom: 0px !important;
 }
 </style>